Update to Xamarin.Forms 4.6.
Enable platform for master and detail page of a master/detail main page. This is now needed to be able to display alerts and arguments from the master page. Since Xamarin.Forms 4.2, Page.Display actions are queued if a page is not platform enabled (https://github.com/xamarin/Xamarin.Forms/pull/7316).